+++ packages/atomic/stencil.config.ts
@@ -39,7 +39,7 @@ export const config: Config = {
   plugins: [
     sass({
       includePaths: ['src/scss/'],
-      injectGlobalPaths: ['src/scss/_global.scss'],
+      injectGlobalPaths: [['src/scss/_global.scss', '*']],
     }),
   ],
   rollupPlugins: {